Allowing Ampute int data

+++ imvc/utils/dataset_utils.py
@@ -52,7 +52,7 @@ class DatasetUtils:
             idxs_to_remove = observed_view_indicator[:,X_idx] == False
             if isinstance(X, pd.DataFrame):
                 X = X.values
-            transformed_X = copy.deepcopy(X)
+            transformed_X = copy.deepcopy(X).astype(float)
             transformed_X[idxs_to_remove, :] = np.nan
             transformed_Xs.append(transformed_X)
         if isinstance(Xs[0], pd.DataFrame):
